+++ studio/src/app/utils/editor/create-slides.utils.tsx
@@ -2,6 +2,8 @@ import {h, JSX} from '@stencil/core';
 
 import {v4 as uuid} from 'uuid';
 
+import {DeckdeckgoPlaygroundTheme} from '@deckdeckgo/slide-playground';
+
 import {SlideAttributes, SlideTemplate} from '../../models/data/slide';
 
 import {EnvironmentDeckDeckGoConfig} from '../../services/core/environment/environment-config';
@@ -47,6 +49,8 @@ export class CreateSlidesUtils {
         resolve(await this.createSlidePoll());
       } else if (template.template === SlideTemplate['ASPECT-RATIO']) {
         resolve(await this.createSlideAspectRatio());
+      } else if (template.template === SlideTemplate.PLAYGROUND) {
+        resolve(await this.createSlidePlayground());
       } else {
         resolve(null);
       }
@@ -271,6 +275,25 @@ export class CreateSlidesUtils {
         <deckgo-slide-youtube key={uuid()} src={src}>
           {title}
         </deckgo-slide-youtube>
+      );
+
+      resolve(slide);
+    });
+  }
+
+  static createSlidePlayground(src: string = undefined, theme: DeckdeckgoPlaygroundTheme = undefined): Promise<JSX.IntrinsicElements> {
+    return new Promise<JSX.IntrinsicElements>((resolve) => {
+      if (!document) {
+        resolve();
+        return;
+      }
+
+      const title = <h1 slot="title"></h1>;
+
+      const slide: JSX.IntrinsicElements = (
+        <deckgo-slide-playground key={uuid()} src={src} theme={theme}>
+          {title}
+        </deckgo-slide-playground>
       );
 
       resolve(slide);