+++ studio/src/global/app.ts
@@ -30,6 +30,7 @@ import '@deckdeckgo/slide-qrcode';
 import '@deckdeckgo/slide-chart';
 import '@deckdeckgo/slide-poll';
 import '@deckdeckgo/slide-aspect-ratio';
+import '@deckdeckgo/slide-playground';
 
 import {setupConfig as setupDeckGoConfig} from '../app/services/core/environment/environment-config';