+++ webcomponents/slides/playground/src/declarations/deckdeckgo-playground-theme.tsx
@@ -1,0 +1,5 @@
+export enum DeckdeckgoPlaygroundTheme {
+  DEFAULT = 'default',
+  LIGHT = 'light',
+  DARK = 'dark',
+}