+++ docs/CHANGELOG.md
@@ -1,3 +1,21 @@
+# 1.5.0 (2020-06-05)
+
+### Features
+
+- new template `playground`
+
+# 1.4.3 (2020-05-31)
+
+### Features
+
+- add strikethrough to `inline-editor`
+
+# 1.4.1 and 1.4.2 (2020-05-29)
+
+### Fix
+
+- QR code slide size calculation
+
 # 1.4.0 (2020-05-21)
 
 ### Features